+++ boxhead/boxhead/boxhead.vcxproj
@@ -165,6 +165,7 @@
     <ClCompile Include="Constants.cpp" />
     <ClCompile Include="Entity.cpp" />
     <ClCompile Include="Framework.cpp" />
+    <ClCompile Include="GameScene.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
@@ -195,6 +196,7 @@
     <ClInclude Include="Constants.hpp" />
     <ClInclude Include="Entity.hpp" />
     <ClInclude Include="Framework.hpp" />
+    <ClInclude Include="GameScene.h" />
     <ClInclude Include="Main.hpp" />
     <ClInclude Include="map.h" />
     <ClInclude Include="pch.hpp" />
