+++ boxhead/boxhead/boxhead.vcxproj.filters
@@ -90,6 +90,9 @@
     <ClCompile Include="Framework.cpp">
       <Filter>소스 파일\Core</Filter>
     </ClCompile>
+    <ClCompile Include="GameScene.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -175,6 +178,9 @@
     </ClInclude>
     <ClInclude Include="zombie.h">
       <Filter>헤더 파일\Contents</Filter>
+    </ClInclude>
+    <ClInclude Include="GameScene.h">
+      <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
