+++ boxhead/boxhead/pch.hpp
@@ -4,4 +4,10 @@
 #ifndef __PCH__
 #define __PCH__
 
+#include <OpenGL.hpp>
+#include <Uniform.hpp>
+
+#include "Constants.hpp"
+#include "Scene.hpp"
+
 #endif
